chore(gas-keys): use u16 for NonceIndex

+++ core/primitives-core/src/types.rs
@@ -20,7 +20,7 @@ pub type StorageUsageChange = i64;
 /// Nonce for transactions.
 pub type Nonce = u64;
 /// Nonce index for gas keys.
-pub type NonceIndex = u32;
+pub type NonceIndex = u16;
 /// Height of the block.
 pub type BlockHeight = u64;
 /// Height of the epoch.
